Lesson 2 Completed

+++ lesson2/week2-ex6.c
@@ -2,29 +2,29 @@
 
 int main() {
     char operation;
-    double n1, n2;
+    float n1, n2;
 
     printf("Enter an operator (+, -, *, /): ");
     scanf("%c", &operation);
 
     printf("Enter two operands: ");
-    scanf("%lf %lf", &n1, &n2);
-    
+    scanf("%f %f", &n1, &n2);
+
     switch (operation) {
         case '+':
-            printf("%.1lf + %.1lf = %.1lf", n1, n2, n1 + n2);
+            printf("%.1f + %.1f = %.1f", n1, n2, n1 + n2);
             break;
 
         case '-':
-            printf("%.1lf - %.1lf = %.1lf", n1, n2, n1 - n2);
+            printf("%.1f - %.1f = %.1f", n1, n2, n1 - n2);
             break;
 
         case '*':
-            printf("%.1lf * %.1lf = %.1lf", n1, n2, n1 * n2);
+            printf("%.1f * %.1f = %.1f", n1, n2, n1 * n2);
             break;
 
         case '/':
-            printf("%.1lf / %.1lf = %.1lf", n1, n2, n1 / n2);
+            printf("%.1f / %.1f = %.1f", n1, n2, n1 / n2);
             break;
 
         default:
